Adds basic page to showcase Angular pipes
Creates a basic page component to demonstrate the usage of built-in Angular pipes such as lowercase, uppercase, titlecase, and date pipes.

Introduces a locale service to handle locale changes and persist the selected locale in local storage. The page includes buttons to change the locale to French, English, or Spanish. A ticking date updates every second, demonstrating the date pipe's reactivity.